% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.r
\name{calulateRFM}
\alias{calulateRFM}
\title{calculateRFM}
\usage{
calulateRFM(data, weight_recency = 1, weight_frequency = 1,
  weight_monetary = 1)
}
\arguments{
\item{data}{- A data.table containing the transaction rececord details for every customer.}

\item{r}{-  Weight of recency.}

\item{f}{-  Weight of frequency.}

\item{m}{-  Weight of monetary.}
}
\value{
Returns a  data.data containing the recency, frequency and monetary
score as well as the weighted final score and the group membership.
}
\description{
Calculate the a weighted RMF score: recency, frequency, and monetary for every customers
}
\details{
\code{data} contains the transactional data. The dataset must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}

